Fixing some values and updating docs
Getting rid of redundant file

removing custom types

trying to restablish src/configs/substrateCustomTypes

removing unnecesary folder

removing unnecesary folder

removing commented code

Main problem solved, looking for refactr

Solving bug and refactoring TransactionStatus componet

stops checking values in useEffefct
