+++ src/hooks/useTransactionNonces.ts
@@ -1,0 +1,92 @@
+// Copyright 2021 Parity Technologies (UK) Ltd.
+// This file is part of Parity Bridges UI.
+//
+// Parity Bridges UI is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Parity Bridges UI is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Parity Bridges UI.  If not, see <http://www.gnu.org/licenses/>.
+
+import { Codec } from '@polkadot/types/types';
+
+import { useEffect, useState } from 'react';
+
+import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
+import useDashboard from './useDashboard';
+import useLaneId from './useLaneId';
+import useLoadingApi from './useLoadingApi';
+import useChainGetters from './useChainGetters';
+import { getTransactionSourceTarget, isTransactionCompleted } from '../util/transactionUtils';
+import { TransanctionStatus } from '../types/transactionTypes';
+import getSubstrateDynamicNames from '../util/getSubstrateDynamicNames';
+interface Props {
+  transaction: TransanctionStatus;
+}
+
+const useTransactionNonces = ({ transaction }: Props) => {
+  const [nonceOfTargetFinalizedBlock, setNonceOfTargetFinalizedBlock] = useState<null | number>(null);
+  const [latestReceivedNonceRuntimeApi, setLatestReceivedNonceRuntimeApi] = useState(0);
+  const { getApiByChain } = useChainGetters();
+
+  const laneId = useLaneId();
+  const areApiLoading = useLoadingApi();
+
+  const { targetTransaction, sourceChain, targetChain } = getTransactionSourceTarget({
+    useSourceTarget,
+    transaction
+  });
+
+  const { bestBlockFinalized: bestBlockFinalizedOnTarget } = useDashboard(targetTransaction);
+
+  const { latestReceivedNonceMethodName } = getSubstrateDynamicNames(sourceChain);
+  const targetApi = getApiByChain(targetChain);
+
+  useEffect(() => {
+    if (!areApiLoading || !transaction || !transaction || isTransactionCompleted(transaction)) {
+      return;
+    }
+    const getNonceByHash = async () => {
+      const blockHash = await targetApi.rpc.chain.getBlockHash(bestBlockFinalizedOnTarget);
+      const latestReceivedNonceCall = await targetApi.rpc.state.call<Codec>(
+        latestReceivedNonceMethodName,
+        laneId,
+        blockHash.toJSON()
+      );
+
+      // @ts-ignore
+      const latestReceivedNonceCallType = targetApi.registry.createType('MessageNonce', latestReceivedNonceCall);
+      const latestReceivedNonce = latestReceivedNonceCallType.toString();
+      setNonceOfTargetFinalizedBlock(parseInt(latestReceivedNonce));
+    };
+
+    const getLatestReceivedNonce = async () => {
+      const latestReceivedNonceCall = await targetApi.rpc.state.call<Codec>(latestReceivedNonceMethodName, laneId);
+      // @ts-ignore
+      const latestReceivedNonceCallType = targetApi.registry.createType('MessageNonce', latestReceivedNonceCall);
+      setLatestReceivedNonceRuntimeApi(parseInt(latestReceivedNonceCallType.toString()));
+    };
+
+    getNonceByHash();
+    getLatestReceivedNonce();
+  }, [
+    areApiLoading,
+    bestBlockFinalizedOnTarget,
+    transaction,
+    laneId,
+    latestReceivedNonceMethodName,
+    targetApi.registry,
+    targetApi.rpc.chain,
+    targetApi.rpc.state
+  ]);
+
+  return { nonceOfTargetFinalizedBlock, latestReceivedNonceRuntimeApi };
+};
+
+export default useTransactionNonces;
